requirements.txt
fastapi>=0.135
uvicorn>=0.41
pandas>=3.0
scikit-learn>=1.8
joblib>=1.5
pytest>=9.0
requests>=2.32
pyyaml>=6.0